allow setting of controller VM/kernel
In general we also want old VMs, like RHEL6. But we (i.e., the LINSTOR
folks) wanted more fine-grained control which distribution is the test
coordinator/controller. They want that because their test suite uses
features only a modern version of Python provides.

Add that, it it wired via the Jenkins config.