vmshed: move sync.WaitGroup out of loop
So far this WG was only used as a per "test group" (i.e., number of VMs)
WG. By moving it out, but still waiting after each inner/per group loop,
the behavior is unchanged. Still, when we relax the fixed waiting after
a group, we can skip the Wait() in the inner loop and just wait for the
final termination. This is the necessary preparation step for it.