vmshed: add caps resolver and add DB caps
So far the only extra capability a distribution could have was to
support ZFS. In the code this was a weird hackery around this "has ZFS
or not". With different types of capabilities like this or that DB type,
we want capabilities generalized. Add this, and add some logic to
(randomly) select sets of VMs that fulfill all the required
capabilities.