initialize virter before executing commands concurrently
This fixes the following problem: vmshed starts multiple VMs
concurrently, and they start successfully, but the private key provided
to the test is only valid for one of them.

It appears that the private key generation on the first run of virter
races. Work around this by running a single virter command before doing
anything concurrently.