add timeout for test image provisioning
We also perform the test image provisioning from the scheduler so that
all errors are collected and output together. This has the added benefit
of making sure we do not use more than the configured number of VMs when
provisioning.

Also move the test timeout to the tests file for consistency and to
avoid having too many command line flags.