add flag to specify RNG seed
It would be helpful to be able to run reproducible tests. This means
that the test and VM sets would be consistent across multiple runs.

This can be accomplished by printing out the random number generator
seed that is being used, and providing a way of overriding that seed via
a command line flag.