restructure and add tags/variants to tests.toml
By now it wasn't possible to match tests to certain VM images, because
sometimes distribution do not provide the same feature set for a test.
e.g.: zfs is missing on centos, or distri X provides a too old version of a tool

"tags": enable you to tag a vmimage with a certain feature and if a test
specify this feature it will only use vmimages with the matching tag.

"variants": variants allows you to run tests with different configurations
and each variant can have multiple variables set that will prepare the test image
in a different way.
e.g. linstor tests run with sql(h2) and etcd as database backend, here only
the setup of the controller is different