do not split --set arguments by comma
With StringSliceVarP, a comma in an argument causes it to be interpreted
as multiple arguments. This makes it difficult to include commas in the
values themselves. Use StringArrayVarP instead.

Without this change, or a similar change in virter, this was necessary:
--set '"""env.FOO=a\,b"""'

Without this change, but with the change in virter, this became:
--set '"env.FOO=a\,b"'

Now with both changes, this works:
--set 'env.FOO=a\,b'

Note that virter still requires the comma to be escaped. However, we no
longer need the additional quotes to get through StringSliceVarP.