Create per-test/provision networks
virter recently learned how to manage virtual networks. We can take advantage
of this feature to:
* Get rid of static DHCP reservation at the start of a test run
* Run VMs for a given test in a fresh virtual network for just that test
* Allow a test to add more network interface (for example: for multipath-tests)

The basic workflow is simple: Before VMs for a test run (or for provisioning)
are started, all needed networks are created. After the VMs stop, the networks
are removed again.

Some care is required around handling subnets assigned to DHCP-enabled
networks. The subnets must be unique across all virtual networks. We
use the suiteState to store which networks are already in-use. This is the
same procedure as used by the virtual machine ids, with some extra code
for iterating subnets.