set boot capacity for vms
Add a new parameters for setting the boot capacity in vms.toml. This allows
setting the capacity of the primary disk, both when provisioning and when
running vms. virter already understood that parameter, but it was
never available for vmshed.

THe original motivation for this is once again openstack: a test suite run
creates over 5GB logs in case of issues. that meant that journald sometimes
dropped the all important logs so that errors could not be traced. Increasing
the / capacity mitigates this issue.