create networks in scheduler
To avoid executing concurrent actions which then have to wait on a
mutex, move the creation of networks to the scheduler. To reduce
unnecessary work, only create as many as we need concurrently. This
means that the networks are no longer uniquely associated to one
provisioning or test run, so a generic network name is used.

Also add DHCP hosts entries when we create the access networks. This
avoids having to ever add DHCP hosts to running networks, which improves
performance significantly.