add "id" field to test run logs
Also implement custom sorting for the standard logger keys so that this
new field is placed before the message. This makes the logs neater and
easier to read.

Note that this special ordering is not applied when logging to a TTY.
Logrus does not allow the order to be overridden in that case.