Windows platform support
In order to support Windows VMs the virter calls have
to be passed the --user Administrator option.

The vms.toml files now have a new field:
    user_name = "Administrator"
should be specified for VMs that are Windows machines.
If this field is omitted the default "root" is
used (so no changes necessary for Linux-only
environments).

As a convenience the --vnc and --vnc-bind-address
options are given for all machines which allows to
debug VMs that fail to start (for example).

The example vms.toml file has been updated to
document these changes.