let users pull VM images from remote location on start
Adds a new option "--pull-template". When specified, every base image
is pulled from the given remote location on first use. The template
given is used to get the remote location used in

  virter image pull <base> <remote>

It accepts a go template string. A user can use "{{ .Image }}" to generically
specify a pull location based on the name of the base image.